Feature/plan rerouting (#22)
* testing plan re-routing - this isn't working yet.

* testing handling with window function

this also is not working, but wanted to keep record of trying it...

* plan updates

plan routing is working now
also added support for checkbox, and sorting the table

* new page ID page

maybe this should be in a table instead of having to paginate? also the saving automatically to the DB is not working.

* new card view!

new card view. also fixed some vercel deploy issues.